+++ src/CommandingOfficers/DefendPeace/GreySky/Cinder.java
@@ -1,6 +1,12 @@
-package CommandingOfficers;
+package CommandingOfficers.DefendPeace.GreySky;
 
 import java.awt.Color;
+
+import CommandingOfficers.Commander;
+import CommandingOfficers.CommanderAbility;
+import CommandingOfficers.CommanderInfo;
+import CommandingOfficers.CommanderAbility.CostBasis;
+import CommandingOfficers.CommanderInfo.InfoPage;
 import Engine.Army;
 import Engine.GameInstance;
 import Engine.GameScenario;
